DozerAI_Code/requirements.txt
# C:\Dozers\DozerAI_Code\requirements.txt

# For loading .env files
python-dotenv~=1.0.1

# Core Supabase client for Python
supabase~=2.4.2

# For FastAPI backend (future days)
fastapi~=0.111.0
uvicorn[standard]~=0.29.0

# For Pydantic (data validation, used by FastAPI and agents)
pydantic~=2.7.1

# For Langfuse observability
langfuse~=2.25.0

# For direct PostgreSQL interaction (used by Day 1 script)
psycopg[binary]~=3.1.18 # For psycopg3

# --- Added for Day 2 ---
# For Google Generative AI (Gemini models)
google-generativeai~=0.5.4 # Check for latest stable

# For Markdown parsing and text splitting
# markdown-it-py for robust Markdown parsing, can be used by custom splitters if needed
# langchain-text-splitters for pre-built advanced splitters
markdown-it-py~=3.0.0
langchain-text-splitters~=0.2.1 # Or latest compatible
langchain-core~=0.2.5 # Often a peer dependency for langchain_text_splitters

# For token counting (often an OpenAI SDK dependency, but good standalone for estimations)
tiktoken~=0.7.0

# --- Added for Day 3 ---
# For LangGraph (Core, and pulls in langchain_core, often LangSmith for tracing if not disabled)
langgraph~=0.0.67 # Or check for latest stable 